Add a VkSwapchainPresentTimingCreateInfoEXT to specify queue size
This new struct extends VkSwapchainCreateInfoKHR, and specifies the size
of the internal queue size holding present timing results. It cannot be
resized during the swapchain's lifetime, as it may involve complex
allocations from the implementation.

A swapchain's internal timing queue fills up when calling vkQueuePresentKHR
if timing collection was enabled previously via vkSetSwapchainTimingEXT. The
queue empties by calling vkGetPastPresentationTimingEXT by the amount of entries
that are returned.

It is made illegal to call vkQueuePresentKHR if the queue is full and
present timing collection is enabled. This change adds a new error enum to
express this, VK_ERROR_PRESENT_TIMING_QUEUE_FULL_EXT.